fixed bugs and refactoring

+++ src/app/todo-item/item.service.ts
@@ -15,7 +15,9 @@ import { AccessType } from '../share/enums/access-type.enum';
 import { DocumentHistoryService } from '../../infrastructure/databases/mongoose/document-history/document-history.service';
 import { ModelsEnum } from '../../models/models.enum';
 import { IDocumentHistory } from '../../infrastructure/databases/mongoose/document-history/interfaces/document-history.interface';
+import { Request } from 'express';
 import { dictionary } from '../../config/dictionary';
+import * as mongoose from 'mongoose';
 @Injectable()
 export class ItemService {
   constructor(
@@ -25,9 +27,20 @@ export class ItemService {
     private readonly docHistoryService: DocumentHistoryService,
   ) {}
 
-  async create(createItemDto: CreateItemDto): Promise<IItem> {
+  async create(req: Request, dto: CreateItemDto): Promise<IItem> {
     try {
-      return await this.itemRepository.create(createItemDto);
+      const currentUser = await this.authService.getUserFromAuthorization(req);
+      if (!dto.userId) {
+        dto.userId = mongoose.Types.ObjectId(currentUser._id);
+      }
+
+      if (currentUser._id !== dto.userId.toString()) {
+        throw new BadRequestException(
+          dictionary.errors.createItemForAnotherUserError,
+        );
+      }
+
+      return await this.itemRepository.create(dto);
     } catch (e) {
       throw new BadRequestException(e);
     }
@@ -42,11 +55,11 @@ export class ItemService {
   }
 
   async getSharedItems(
-    userId: string,
+    req: Request,
     pagination: PaginationOptions,
   ): Promise<PaginateResult<IItem>> {
     const sharedItemsId: string[] = [];
-    const data = await this.shareService.getSharedItemByUser(userId);
+    const data = await this.shareService.getSharedItemByUser(req, pagination);
     data.docs.forEach(sharedItem => {
       sharedItemsId.push(sharedItem.itemId);
     });
@@ -56,7 +69,7 @@ export class ItemService {
         {
           _id: sharedItemsId,
         },
-        pagination,
+        {},
       );
     } catch (e) {
       throw new InternalServerErrorException(e);
@@ -64,26 +77,23 @@ export class ItemService {
   }
 
   async getByUser(
-    token: string,
-    userId: string,
+    req: Request,
     pagination: PaginationOptions,
   ): Promise<PaginateResult<IItem>> {
     try {
-      const currentUser = await this.authService.getCurrentUser(token);
-      if (userId !== currentUser._id) {
-        throw new ForbiddenException(
-          dictionary.errors.gettingItemsAnotherUserError,
-        );
-      }
-
-      return this.itemRepository.getByUser(userId, pagination);
+      const currentUser = await this.authService.getUserFromAuthorization(req);
+      return this.itemRepository.getByUser(currentUser._id, pagination);
     } catch (e) {
       throw new BadRequestException(e);
     }
   }
 
-  async update(token, _id: string, payload: Partial<IItem>): Promise<IItem> {
-    const currentUser = await this.authService.getCurrentUser(token);
+  async update(
+    req: Request,
+    _id: string,
+    payload: Partial<IItem>,
+  ): Promise<IItem> {
+    const currentUser = await this.authService.getUserFromAuthorization(req);
     const currentItem = await this.itemRepository.findById(_id);
 
     if (
@@ -109,11 +119,19 @@ export class ItemService {
       }
     }
 
-    throw new ForbiddenException(dictionary.errors.permissionsForUpdateError);
+    throw new ForbiddenException(dictionary.errors.permissionForModifyError);
   }
 
-  async delete(_id: string): Promise<any> {
+  async delete(req: Request, _id: string): Promise<any> {
     try {
+      const currentUser = await this.authService.getUserFromAuthorization(req);
+      const currentItem = await this.itemRepository.findById(_id);
+      if (currentUser._id !== currentItem.userId.toString()) {
+        throw new ForbiddenException(
+          dictionary.errors.permissionForModifyError,
+        );
+      }
+
       return this.itemRepository.delete(_id);
     } catch (e) {
       throw new InternalServerErrorException(e);
@@ -125,7 +143,7 @@ export class ItemService {
     options: PaginationOptions,
   ): Promise<PaginateResult<IDocumentHistory>> {
     try {
-      return await this.docHistoryService.getHistoryByDocId(id, options);
+      return this.docHistoryService.getHistoryByDocId(id, options);
     } catch (e) {
       throw new InternalServerErrorException(e);
     }
